Move setFetchError and setIsLoading to their own slice 'fetchStatusSlice'
